# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  "/api/v1/user/cards":
    post:
      tags:
        - User Cards Management
      summary: Get user cards list. (YES, it's POST)
      description: Returns a paginated list of cards belonging to the currently authenticated
        user. The user identifier is extracted from the JWT token. Requires **ROLE_USER**
        authority.
      operationId: getMyCards
      parameters:
        - name: pageable
          in: query
          description: Pagination parameters including page, size, and sort.
          required: true
          schema:
            type: object
          example:
            page: 0
            size: 10
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetCardsRequestDto"
        required: true
      responses:
        '200':
          description: Card list successfully retrieved.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/GetCardsResponseDto"
        '400':
          description: Invalid request parameters provided.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required permissions.
      security:
        - BearerAuth: []
  "/api/v1/user/cards/transfer":
    post:
      tags:
        - User Cards Management
      summary: Transfer between own cards.
      description: Executes a fund transfer from one card account of the current user
        to another card account of the same user. Requires **ROLE_USER** authority
        and ownership of both cards.
      operationId: transferBetweenOwnCards
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TransferRequestDto"
        required: true
      responses:
        '200':
          description: Transfer successfully completed.
        '400':
          description: Insufficient funds or invalid transfer parameters.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required permissions.
        '404':
          description: One or both cards not found.
      security:
        - BearerAuth: []
  "/api/v1/user/cards/block":
    post:
      tags:
        - User Cards Management
      summary: Block card.
      description: Submits a request to block the specified card of the current user.
        Once blocked, transactions on this card will be unavailable. Requires **ROLE_USER**
        authority and ownership of the card.
      operationId: blockCard
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/BlockCardRequestDto"
        required: true
      responses:
        '200':
          description: Card successfully blocked.
        '400':
          description: Data validation error occurred.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required permissions.
        '404':
          description: Card not found.
      security:
        - BearerAuth: []
  "/api/v1/user/cards/balance":
    post:
      tags:
        - User Cards Management
      summary: Get card balance. (YES, it's POST)
      description: Returns the current balance and currency for the specified card
        belonging to the user. Requires **ROLE_USER** authority and ownership of the
        card.
      operationId: getBalance
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetCardBalanceRequestDto"
        required: true
      responses:
        '200':
          description: Card balance successfully retrieved.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/GetCardBalanceResponseDto"
        '400':
          description: Invalid card ID format or empty request body.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required permissions.
        '404':
          description: Card not found or does not belong to the user.
      security:
        - BearerAuth: []
...
  "/api/v1/admin/cards/search":
    post:
      tags:
        - Admin Cards Management
      summary: Get filtered list of all cards. (YES, it's POST)
      description: Returns a paginated list of bank cards matching the search criteria
        specified in the request body. Requires **ROLE_ADMIN** authority.
      operationId: getAllCards
      parameters:
        - name: pageable
          in: query
          description: Pagination parameters including page, size, and sort.
          required: true
          schema:
            type: object
          example:
            page: 0
            size: 10
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GetCardsAdminRequestDto"
        required: true
      responses:
        '200':
          description: Card list successfully retrieved.
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/GetCardsResponseDto"
        '400':
          description: Invalid search or pagination parameters provided.
        '401':
          description: User is not authenticated.
        '403':
          description: User does not have the required administrative permissions.
      security:
        - BearerAuth: []
...
components:
  schemas:
    GetCardsRequestDto:
      type: object
      description: Request payload used to fetch and filter a list of bank cards belonging
        to a specific user.
      properties:
        userId:
          type:
            - string
            - 'null'
          format: uuid
          description: The unique identifier of the user whose cards are being requested.
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type:
            - string
            - 'null'
          description: Optional card status filter. If omitted or null, all cards
            for the user will be returned.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
    GetCardResponseDto:
      type: object
      description: Response payload containing complete information and current operational
        state of a bank card.
      properties:
        uuid:
          type: string
          format: uuid
          description: The unique business identifier (UUID) assigned to the bank
            card.
          example: e4219747-f42d-4e32-92a8-e67d9841910e
        cardNumber:
          type: string
          description: The masked or encrypted credit card number format for secure
            visualization.
          example: 4444 44** **** 4444
        holderName:
          type: string
          description: The full name of the card owner as printed on the physical
            card.
          example: JOHN DOE
        expiryDate:
          type: string
          format: date-time
          description: The exact expiration date and time of the card in UTC ISO-8601
            format.
          example: '2029-12-31T23:59:59Z'
        status:
          type: string
          description: The current lifecycle and operational status of the bank card.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
        balance:
          type: number
          description: The current available financial balance amount remaining on
            the card.
          example: 15450.5
    GetCardsResponseDto:
      type: object
      description: Response payload containing the paginated collection of retrieved
        bank cards along with execution metadata.
      properties:
        cards:
          type: array
          description: The list of bank card details matching the requested filter
            criteria or belonging to the target user.
          items:
            "$ref": "#/components/schemas/GetCardResponseDto"
        pageNumber:
          type: integer
          format: int32
          description: The current page index position within the paginated record
            set, starting from index 0.
          example: 0
        pageSize:
          type: integer
          format: int32
          description: The maximum number of card records requested or allowed to
            be displayed on a single page slice.
          example: 10
        totalElements:
          type: integer
          format: int64
          description: The absolute total count of all bank card records existing
            in the database that match the criteria.
          example: 47
        totalPages:
          type: integer
          format: int32
          description: The calculated total count of available pages based on total
            elements and the requested page size.
          example: 5
    TransferRequestDto:
      type: object
      description: Request payload used to execute a financial transfer between two
        bank cards.
      properties:
        userId:
          type: string
          format: uuid
          description: The unique identifier of the user executing the transaction.
          example: 123e4567-e89b-12d3-a456-426614174000
        fromCardId:
          type: string
          format: uuid
          description: The unique identifier of the source bank card from which funds
            will be debited.
          example: 987f6543-e21b-78d3-c987-123456789abc
        toCardId:
          type: string
          format: uuid
          description: The unique identifier of the destination bank card to which
            funds will be credited.
          example: 550e8400-e29b-41d4-a716-446655440000
        amount:
          type: number
          description: The financial amount to be transferred. Must be a positive
            value.
          example: 250.5
          minimum: 0.01
      required:
        - amount
        - fromCardId
        - toCardId
        - userId
    BlockCardRequestDto:
      type: object
      description: Request payload used to block a specific bank card within the system.
      properties:
        userId:
          type:
            - string
            - 'null'
          format: uuid
          description: The unique identifier of the user who owns the card. This field
            is optional.
          example: 123e4567-e89b-12d3-a456-426614174000
        cardId:
          type: string
          format: uuid
          description: The unique identifier of the bank card that needs to be blocked.
          example: 987f6543-e21b-78d3-c987-123456789abc
      required:
        - cardId
    GetCardBalanceRequestDto:
      type: object
      description: Request payload used to check the available financial balance of
        a specific bank card.
      properties:
        userId:
          type:
            - string
            - 'null'
          format: uuid
          description: The unique identifier of the user who owns the card. This field
            is optional.
          example: 123e4567-e89b-12d3-a456-426614174000
        cardId:
          type: string
          format: uuid
          description: The unique identifier of the bank card whose balance is being
            requested.
          example: 987f6543-e21b-78d3-c987-123456789abc
      required:
        - cardId
    GetCardBalanceResponseDto:
      type: object
      description: Response payload containing the current financial balance of the
        requested bank card.
      properties:
        balance:
          type: number
          description: The current available financial balance amount on the card.
          example: 1250.75
...
    GetCardsAdminRequestDto:
      type: object
      description: Request payload used to fetch and filter a list of bank cards belonging
        to a specific user.
      properties:
        userId:
          type:
            - string
            - 'null'
          format: uuid
          description: The unique identifier of the user who owns the card. This field
            is optional.
          example: 123e4567-e89b-12d3-a456-426614174000
        status:
          type:
            - string
            - 'null'
          description: Optional card status filter. If omitted or null, all cards
            for the user will be returned.
          enum:
            - ACTIVE
            - BLOCKED
            - EXPIRED
          example: ACTIVE
...